refactor rich text charater limit example block
