+++ example/src/blocks/rich-text-character-limit/index.js
@@ -1,50 +1,9 @@
 import { registerBlockType } from '@wordpress/blocks';
-import { __ } from '@wordpress/i18n';
-import { useBlockProps } from '@wordpress/block-editor';
-import { RichTextCharacterLimit } from '@10up/block-components';
 
-const NAMESPACE = 'example';
+import metadata from './block.json';
+import { BlockEdit } from './edit';
 
-registerBlockType( `${ NAMESPACE }/rich-text-character-limit`, {
-	apiVersion: 2,
-	title: __( 'Rich Text Character Limit', NAMESPACE ),
-	description: __( 'Example Block to show the Rich Text Character Limit in usage', NAMESPACE ),
-	icon: 'smiley',
-	category: 'common',
-	example: {},
-	supports: {
-		html: false
-	},
-	attributes: {
-		title: {
-			type: 'string',
-			default: ''
-		}
-	},
-	edit: (props) => {
-		const {
-			attributes: { title },
-			setAttributes
-		} = props;
-
-		const blockProps = useBlockProps();
-
-		return (
-			<div {...blockProps}>
-				<RichTextCharacterLimit
-					limit={30}
-					enforce={true}
-					tagName="h2"
-					value={title}
-					placeholder={ __( 'Enter some text', NAMESPACE ) }
-					onChange={(title) => setAttributes({title})}
-					allowedFormats={[
-						'core/bold',
-						'core/link'
-					]}
-				/>
-			</div>
-		)
-	},
+registerBlockType(metadata, {
+	edit: BlockEdit,
 	save: () => null
-} );
+});
